# docs/docs_swagger/swagger.yaml
basePath: /api
...
host: localhost:10031
info:
  contact:
    email: support@yourcompany.com
    name: API Support
    url: http://www.yourcompany.com/support
  description: 智能门禁管理系统API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ILock HTTP Service API
  version: "1.0"
paths:
...
  /admin/device-building/bind:
    post:
      consumes:
      - application/json
      description: 将一个或多个设备绑定到一个建筑物
      parameters:
      - description: 绑定信息
        in: body
        name: bindInfo
        required: true
        schema:
          type: object
      - description: 设备ID列表
        in: body
        name: deviceIds
        required: true
        schema:
          items:
            type: integer
          type: array
      - description: 建筑ID
        in: body
        name: buildingId
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 绑定成功消息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 输入参数错误或绑定失败
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 绑定设备到建筑
      tags:
      - Device-Building
  /admin/device-building/building:
    get:
      consumes:
      - application/json
      description: 根据设备ID获取其所属的建筑物信息
      parameters:
      - description: 设备ID
        in: query
        name: deviceId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 建筑信息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 无效的设备ID或获取建筑失败
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取设备所属建筑
      tags:
      - Device-Building
  /admin/device-building/devices:
    get:
      consumes:
      - application/json
      description: 获取与指定建筑物关联的所有设备信息及其状态
      parameters:
      - description: 建筑ID
        in: query
        name: buildingId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 设备列表信息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 无效的建筑ID或获取设备失败
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取建筑绑定的设备
      tags:
      - Device-Building
  /admin/device-building/unbind:
    post:
      consumes:
      - application/json
      description: 解除一个设备与其关联的建筑物的绑定关系
      parameters:
      - description: 解绑信息
        in: body
        name: unbindInfo
        required: true
        schema:
          type: object
      - description: 设备ID
        in: body
        name: deviceId
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 解绑成功消息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 输入参数错误或解绑失败
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 解绑设备与建筑
      tags:
      - Device-Building
...
  /admin/file-advertisement/advertisement:
    get:
      consumes:
      - application/json
      description: 获取与指定文件关联的广告信息
      parameters:
      - description: 文件ID
        in: query
        name: fileId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 广告信息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 无效的文件ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 根据文件获取广告
      tags:
      - File-Advertisement
  /admin/file-advertisement/bind:
    post:
      consumes:
      - application/json
      description: 将一个文件绑定到一个广告
      parameters:
      - description: 绑定信息
        in: body
        name: bindInfo
        required: true
        schema:
          type: object
      - description: 广告ID
        in: body
        name: advertisementId
        required: true
        schema:
          type: integer
      - description: 文件ID
        in: body
        name: fileId
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 绑定成功消息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 输入参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 广告或文件不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 绑定文件到广告
      tags:
      - File-Advertisement
  /admin/file-advertisement/file:
    get:
      consumes:
      - application/json
      description: 获取与指定广告关联的文件信息
      parameters:
      - description: 广告ID
        in: query
        name: advertisementId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 文件信息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 无效的广告ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 根据广告获取文件
      tags:
      - File-Advertisement
  /admin/file-advertisement/unbind:
    get:
      consumes:
      - application/json
      description: 解除一个广告与其绑定的文件的关系
      parameters:
      - description: 广告ID
        in: query
        name: advertisementId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 解绑成功消息
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 无效的广告ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 解绑文件与广告
      tags:
      - File-Advertisement
...
securityDefinitions:
  BearerAuth:
    description: 输入带有Bearer前缀的用户认证token
    in: header
    name: Authorization
    type: apiKey
  DeviceAuth:
    description: 输入带有Bearer前缀的设备认证token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"